+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.15.0] - (14.01.2023)
+## Changes
+- added Channel Motion Sensor for use with automations (every channel change report motion)
+- config.schema updated
+
 ## [3.14.9] - (04.01.2023)
 ## Changes
 - fix save target visibility